Use shorter scale specifications
When working on #41693 I totally forgot that I'd added shorter names for
the various coord::scale values to use in project_to etc. without being
so verbose.

This makes the API more consistent because the names of the scales match
the abbreviations in the type names.

Convert to the shorter names.

Strictly a find-and-replace commit, so hopefully should be safe.

(cherry picked from commit ab7a370c5055857c7c817a6815cf2e5086b6dfa9)