Port over some DDA relic effects to BN
1. Changes to character.cpp adding code that looks for MAX_HP and
REGEN_HP enchantments. Did the MAX_HP change in `Character::recalc_hp`
instead, as the method used in the PR is rooted in the unhardcoded
bodyparts change.
2. Change to character.cpp implementing the REGEN_STAMINA value.
3. Change to character.cpp implementing the FATIGUE value.

PRs being ported over:
1. https://github.com/CleverRaven/Cataclysm-DDA/pull/42030
2. Part of https://github.com/CleverRaven/Cataclysm-DDA/pull/42090

Properties that will be a priority for my to port over next PR:
1. METABOLISM and THIRST, since I ported over FATIGUE. Since BN doesn't
have the metabolic rate stuff, from what I can tell I'll need to put the
implementation in Consumption.cpp instead for METABOLISM to work. We'll
see.
2. MAX_STAMINA, since I ported over REGEN_STAMINA.

Side note, I was trying to find SOCIAL_PERSUADE since it's also one
Arcana uses, but checked on
https://github.com/CleverRaven/Cataclysm-DDA/issues/33837 and realized
it actually hasn't been implemented yet. Not important to me, Arcana's
use of it is basically just a flavor effect, I'll just have to remember
to remove it from my mod JSON so I don't get confused in the future.