Set enchantment modifiers to not round floats
And this should fix it, as it did once I got the tests running. Turns
out the boolean was the issue after all. Was advised to make them not
round anytime it's working with a float, and it seems the max HP one is
the only one that uses a straight int.