Show presence of monsters when examining
When using `e` to examine an adjacent tile, the existence of monsters
(hostile, friendly, or neutral) was not indicated in any way.

Change `e` to print "There is a zombie" or "There is a cow" for example,
when such a creature is in the examined space.

(cherry picked from commit 7b4a4e44f13fe6692190543f46978ab75575d00b)