Pipe weapons easier to craft
The pipe rifle (.22, 9mm) and the pipe shotguns (both single and double) have had their recipes changed to require fabrication 2 and mechanics 1. They also no longer require a hacksaw - a multitool will suffice.
The purpose of this change is to make them easily craftable at the point in the game where they're useful - at the very beginning when you may have loose ammo from killing a zombie, but no gun to fire it.