Parametric mapgen part 1: global random terrain selection (#48529)
* Allow access to JsonArray members as JsonValue

* Create load_weighted_list function

This is a generic function for loading a weighted_list from json that
can be used to reduce code duplication and ensure json consistency.

* Remove undefined declaration

This function had no definition and no callers.

* Implement "parameters" for mapgen

This is an initial implementation of parameters for mapgen, which
currently has those parameters act more like variables than parameters.

Essentially all it currently permits is to choose a terrain at random
and use that in a palette.  Thus choosing a thing at random which
applies to an entire map, rather than choosing at random on a per-tile
basis.

This is the first step towards true parameters for mapgen; hence the
currently-misleading name.

* Add docs for new parameters feature.

* Minor grammar/style fixes

Co-authored-by: actual-nh <74678550+actual-nh@users.noreply.github.com>

Co-authored-by: actual-nh <74678550+actual-nh@users.noreply.github.com>