Mapgen params to map special scope and palettes
Previously mapgen parameters only applied to a single OMT, and could not
be in palettes.  Remove both those limitations.

Now a parameter can have map_special scope (indeed, that's the default)
so that the same value is chosen consistently across any specific
overmap_special.

Moreover, parameters can be placed in palettes, and are automatically
extracted from all the palettes relevant to a particular piece of
mapgen.