Ensure mapgen order is sensible (#50597)
When terrain and furniture were merged into the same mapgen system as
everything else it changed the order in which mapgen happened.  In
particular, some terrain was now placed after nested mapgen, which
caused the terrain from the nest to be overridden.

Add the ability to define an ordering in which the jmapgen_pieces should
be applied, and ensure that terrain and furniture go first.  This should
more closely match what happened before the changes in #48498.

In particular, this fixes #49788, the issue with missing walls in
basements.