Add Boxing to Self-Defense Classes trait
Recreates https://github.com/CleverRaven/Cataclysm-DDA/pull/50243 for BN. Basically, Boxing can't be selected at character generation save by two separate professions (one of which is only in the later versions of Magiclysm). As I could literally go up the street and enroll in a boxing gym, this doesn't make a lot of sense and it's less fun to boot.