Tool belt rigid and less encumbering
Tool belt assumes 22 max encumbrance which translates to roughly 11 encumbrance when 6 items are in the belt which is simply too much for the benefit it provides. There are much better alternatives for storing tool and this slot (torso waist) has a very strong opponent which is survivor harness.
Survivor utility belt which contains all necessary tools has only 2 encumbrance and provides 0.5 L volume on top of it.
I find this change making this item at least marginally useful other than a crafting component for crafting other better items.