Give AISLE flag to internal doors (#1236)
* Give AISLE flag to internal and wood doors

As per suggestion, internal doors should have a stepable aisle flooring integrated to easily maneuver inside a vehicle. This reasoning has been taken to wooden doors, which would require a frame to mount into a car in the first place.

* revert wooden door changes

Removal of the "AISLE" flag from wooden doors, as this would be a change requiring it's own discussion and balance changes.