Add a star fort with working cannon and zombie reenactors
Due to not being able to get cherry-pick to work I used diff and copy paste to add the changes. The result is all the commits of DDA PR #45215 being squashed into this one. I made sure to preserve the authors, but this does result in the exact details of which contributions were whos being lost in the BN record.

Co-Authored-By: jokermatt999 <jokermatt999@gmail.com>
Co-Authored-By: Mark Langsdorf <mlangsdo@redhat.com>