Remove handload quality distinction (#1481)
* Make Factory Ammo

* Obsoletion, Recipe balancing, Migration Test

Uncraft and Craft were heavily unbalanced and pulling materials out of thin air during uncraft.

Moved all reloaded ammo into the obsoletion/items.json

MIgration missed 5x50mm flechettes, fixed that.

* Fix bugs

Bugtest revealed human error. Fixed and retested.

* Styling fixes

* Fixed Item groups. reloaded_4570_bp unobsoleted.

reloaded_4570_bp probably removed because of reloaded prefix, but is actually a blackpowder load, and the recipe still remains in the game, so I put it back since it appears to be an oversight.

All reference to reloaded ammunition in item_groups has been replaced with factory ammo.

* 4570_bp

For some reason this predates my changes to the system???

This means that all this time we've had a recipe that leads to an obsoleted item.

Added back into the game and made it reversible because I don't really understand why it was removed.