Fix crossbow gunmod compatibility (#1526)
* Fix crossbow gunmod compatibility

So recently I was poking through item.cpp and was reminded that the gunmod function that separates crossbows from other types of ranged weapon was never updated to account for crossbows using rifle skill, as it was switched from archery by https://github.com/CleverRaven/Cataclysm-DDA/pull/21545

On looking at the function I decided that there were two ways to fix it. Either convert the "if ammo is bolt or it's a bullet crossbow, it's a crossbow" statement to not care about the archery skill, since it triggering would make the function return before it could grab the gun skill anyway. Or I could implement a flag check.

The main reason to do this is, aside from checking for specific ammo types and item IDs not being a good way to do this, there are going to be items that fire bolts but shouldn't count as a crossbow, example being the pneumatic bolt driver, plus you may have items that should be a crossbow but wouldn't be under that check. No vanilla examples of the former come to mind, though outside repo of course adds the example of wraithslayer crossbows in Arcana.

* Update item.cpp