Hopefully final C++ skeleton
COMPACT flag to make power armor mods not cause layering penalties.

Changes to logic so that Externals can only be worn on parts not already covered by an exoskeleton.

Changes hauling frame to be a mod instead of an external in light of this.

Changes to logic so that mods need an external or exoskeleton to attach to.