Humvee GL ammo spawn change for Aftershock
Grenade launcher in Humvee in Aftershock was set to use 40x46mm ammo. The correct ammo caliber is 40x53mm. 46mm is for handheld launchers, 53mm is for the vehicle based GL. The 46mm in Humvee always caused a debug error message, which you may have seen a few times, and immediately forgotten about it. I also removed the EMP grenade spawn possibility because the grenade type doesn't exist for 53mm. Feel free to create it.