Make path appear in __FILE__ for tests
The DebugLog printing of filenames relies on __FILE__.  This is based on
the filename passed to the compiler.  For tests, that lacked a directory
name.  So, __FILE__ lacked the context for the filename, which meant the
GitHub problem matcher couldn't find the file causing the error.

Add '../tests/' to the filename passed to the compiler to work around
this.

Co-authored-by: John Bytheway <jbytheway@gmail.com>