Nonconductive swords with wooden handles (#2866)
* Basic first start

Swords crafted using wood now are nonconductive, since they have wooden handles. Also added a basic dagger weapon which is basically a slightly better baselard. If you can forge a broadsword, you should be able to forge a well-made dagger.

* Removed rapid from dagger

The rapid technique is apparently not very good due to its lower damage, so I removed it.

* Made dagger autolearn again

It's a dagger, not a fancy sword. You can learn to make a proper dagger with enough skill.

* Fixed textbook issue