Fixes for the balance fixes
Reduced 00 shot damage down to 80 and birdshot to 50, and reverted cone length for birdshot and dragon's breath back to 8. Also adjusted the range and damage of the 20x66mm shot to have more range and the .410 shot to have the same range as the 00 shot, and the damage of both now has the same ratio as it did before the previous buff. Paper birdshot was buffed to match the same previous damage ratio, and flintlock ammo was given a range buff as well. The shot had a range of 4 before now, wtf.