Lightened turbines (#3396)
* PR #3383 - Rebalancing of gas turbine

Change of the 3 gas turbine weight to be consistent with their real life equivalent. They now live up to their description of having a good power to weight ratio.

* Add files via upload

Everything is better with commas.

* Unit change for turbine weights

changed from grams to kilograms. Might be a good idea to do the same for the rest of the engines...