refactor(port): tidy specials data and mutable connections (#3834)
* Refactor overmap special data (#52234)

Previously the two types of overmap special (fixed vs mutable) stored
their data completely differently.  The fixed data was inline, and the
mutable data was via a pointer.  And all the functions useing those had
to switch on the type to decide which to use.

Refactor the code so that both types of special are instead derived
classes from an abstract overmap_special_data class.

This helps clarify some situations that were not handled correctly in
the mutable case.  Mapgen arguments were not being gathered from mutable
special overmaps, the dimensions were not being calculated properly, and
they were not respecting the must_be_unexplored flag.  The first two of
these issues are now fixed; the last remains (but there's now a TODO
about it).

Co-authored-by: Kevin Granade <kevin.granade@gmail.com>

* Add connections support to mutable specials

Each mutable_overmap_terrain in a mutable overmap special can now
define connections which will be routed back to the nearest city or rod
in the same way that regular overmap special connections are.

Demonstrate this feature using the mutable mine.

* Document mutable overmap special connections

* style(autofix.ci): automated formatting

---------

Co-authored-by: John Bytheway <52664+jbytheway@users.noreply.github.com>
Co-authored-by: Kevin Granade <kevin.granade@gmail.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>