feat(balance): Steel bowstring, repeating crossbow debuff (#4242)
* Steel crossbow string debuff

Made the damage boost 5 instead of 15 and added a reload modifier of 15. Also changed the description to reflect this.

* Repeating crossbow debuff

Made the repeating crossbow have a slight damage reduction and range reduction to reflect its greater fire rate, making the standard crossbow still have a utility.

* Renamed steel bowstring

On the suggestion from Scarf, the steel bowstring is now the steel prod, a steel reinforcement for the prod (or bowlimbs) of a crossbow.