feat: extend add_spawn() to handle creating pets
So far map::add_spawn() and the underlying spawn_point struct have been
operating on the binary notion that a spawn can be friendly or not. This
PR extends those APIs to also support generating monsters that are pets
of the player, while also honoring backwards compatibility with parts of
the codebase that still want to operate on the binary friendly notion,
such as the save/load system for submaps.