fix: dont spawn eggs with negative age values
This is a fix for the corner-case where a monster that lays eggs is
forced to reproduce via the debug menu. In such a case, the age of the
spawned egg would be negative as the birth-timer points to a point in
time in the future for the next valid reprouce check of the monster.

In this corner-case, we just use the current-turn timestamp for the
newly spawned egg.