feat: prevent chickens from escaping chicken coop
The idea is to prevent pet chickens from escaping the chicken coop when
the player goes in/out of the coop by adding a specific flag both
to the open chicken gate tile as well as the chickens themselves.

This flag will prevent the pet chicken from considering moving to the
chicken gate tile a valid move. This concept can be applied to other
types of gates and animals by reusing the new flags PASTURE_GATE and
PET_AVOID_PASTURE_GATE.