fix: nets should not deal damage
As the title implies, nets should not be dealing damage to the monsters
they envelop. No matter how hard you throw a net, you are just throwing
a net, it shouldn't kill a chicken in 2 throws.

Introduce NO_DAMAGE flag to deal with projectiles that should apply
effects but not actually deal damage using typical logic.