Added thermoblade
Added in thermoblade, a utility knife that heats itself using an atomic battery. It can show up almost anywhere a hunting knife can (at a lower chance), except on police since it's legally questionable. It can be crafted yourself using a knife, electronics supplies, and a plutonium cell if you have 6 electronics and a lab recipe book or 8 electronics otherwise.