feat(content): thermoblade (#4500)
* Added thermoblade

Added in thermoblade, a utility knife that heats itself using an atomic battery. It can show up almost anywhere a hunting knife can (at a lower chance), except on police since it's legally questionable. It can be crafted yourself using a knife, electronics supplies, and a plutonium cell if you have 6 electronics and a lab recipe book or 8 electronics otherwise.

* Removed durable flag

The thermoblade is no longer durable, since this version is a handmade grey-market variety while the Subnautica one is 3D printed.

* style(autofix.ci): automated formatting

* Removed technique, added FLAMING

Removed the thermal stab technique and the bonus damage and gave the thermoblade the FLAMING flag instead.

* Fixed recipebook learn level

* style(autofix.ci): automated formatting

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>