Whips
The bullwhip and scourge have been resized and reweighted to be closer to an IRL whip of that size, allowing them to serve as lightweight but relatively low damage and low durability long-range melee weapons. The bullwhip also now has the belt-clip flag, since you could easily just loosely tie one into a slipknot to keep it in place. The crafting time is also lowered to be in-line with BN's generally reduced crafting times.