feat(balance): whip rebalancing (#4510)
* Whips

The bullwhip and scourge have been resized and reweighted to be closer to an IRL whip of that size, allowing them to serve as lightweight but relatively low damage and low durability long-range melee weapons. The bullwhip also now has the belt-clip flag, since you could easily just loosely tie one into a slipknot to keep it in place. The crafting time is also lowered to be in-line with BN's generally reduced crafting times.

* Removed rope requirement

Bullwhips are almost always entirely leather, with multi-layered whips just using more leather on the inside; I'm not even sure why the crafting recipe required rope to begin with.