feat(balance): toolboxes are not mechanical (#4512)
* Toolbox No Mechanics

The toolbox no longer needs mechanics to build, since it is literally just storing a bunch of tools inside of a metal box held together with nails and wire. It now instead requires a level 2 hammering tool.

* No hammering

The hammer is already a tool in the recipe, no need to require another one.