feat(content): additional unarmed weapons (#4530)
* Initial commit

This branch adds late game additions to unarmed weapons in the form of a jamadhar, a double-bladed variant of the katar weapon, excellent for piercing, as well as an iteration of the impact knuckles existing in DDA into BN.

* small update

bit of streamlining after testing around with various parameters

* small updoot

more stat streamlining

* more streamlining

bleh

* style(autofix.ci): automated formatting

* Update data/json/items/melee/unarmed_weapons.json

* rename of jamadhar id, adding item spawn rate

As per suggestion, renamed the ID to resemble the resulting name to be as similar as possible. Also added them to  the appropriate antique tables with probability. Can't figure out a place for impact knuckles though, as that has no actual basis in reality or the setting.

* Update data/json/items/melee/unarmed_weapons.json

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Chaosvolt <chaosvolt@users.noreply.github.com>