feat(balance): Unarmed weapon adjustments (#4543)
* Knuckles + cestus buff

Brass knuckles, steel knuckles, and the cestus have been buffed, now doing 9, 8, and 9 damage each. If the nail knuckles can deal 3 bash and 4 cut, then a piece of actual metal attached to your fists can do more than that.

* Sheaths, belt clips

Made the sai use the knife sheath and the cestus use the belt clip, since sai are basically pointed but not edged blades and cestuses are just leather gloves with steel knuckles and could easily be tied to your belt.

* Steel knuckles belt clip

* Fixed recipes

Corrected the recipes for the knuckle katar, cestus, and steel knuckles to use unarmed skill instead of bashing to learn and craft.