feat(balance): Survivor helmet buff (#4549)
* Survivor helmet buff

Survivor helmets were generally inferior to the army helmet despite them all having the army helmet as an ingredient. This is now fixed.

* Increased thickness for tactical helm

Increased thickness and changed material of kevlar to rigid kevlar for tactical helmet and tactical full helmet to match the army helmet, and increased survivor helmet's thickness by one.

* Tac helmet survivor helmet

The tactical helmet can be used to create  the survivor helmet, but not the heavy or superalloy survivor helmets. This is now corrected.