Basic Implementation for vehicle damage world option
This adds a new world option to modify vehicle damage. The purpose is to have a global value to alter initial vehicle health.
Vehicles may be more broken with a value > 1 than usual so players may need to scrounge up vehicle parts to get a working one.
In comparison the No Hope mod modifies most mapgen JSON files to create wrecks instead of vehicles.

No Hope does it by setting values for "number" such as [ 3, 4 ] inside vehicle_spawn.json
The problem of this solution is that whenver a new map is created that uses "place_vehicles" it would need to be modified. This is hard to maintain and changes to the maps need to be backported.

The multiplication inside this commit is only a very basic solution that may be changed at will. There were no considerations towards the distribution of non-damaged, damaged and broken components as well as the set_hp values inside the following function