Spear size/weight rebalance
Made spears have consistent size and weights, with variances based on material, to ensure that their DPS is more even. Made it so that only pikes have the "ALWAYS_TWOHAND" flag. Made the survivor naginata not have a to-hit penalty and instead gave it a +1 to hit like the pipe spear.