Spiked club, minor weapon tweaks
Added spiked club item, a wooden club with bands of iron and spikes crafted onto it. Spiked clubs were a common medieval weapon among peasants. Also changed the morningstar and warhammer to have +1 to hit, instead of -1 and 0 respectively, since morningstars are literally just heavily spiked maces and warhammers are just hammers meant to be used as a weapon.