feat(content): Spiked club, minor weapon tweaks (#5083)
* Spiked club, minor weapon tweaks

Added spiked club item, a wooden club with bands of iron and spikes crafted onto it. Spiked clubs were a common medieval weapon among peasants. Also changed the morningstar and warhammer to have +1 to hit, instead of -1 and 0 respectively, since morningstars are literally just heavily spiked maces and warhammers are just hammers meant to be used as a weapon.

* Added metal sawing requirement

* style(autofix.ci): automated formatting

* Update data/json/items/melee/bludgeons.json

* Update data/json/items/melee/bludgeons.json

* Readjusted damage

* Changed wood requirements

The spiked mace now requires a log, shillelagh, or baseball bat to craft, as befitting a more complex hardwood weapon.

* Changed looks-like and damage

Bumped up damage to compensate for new hardwood wood type, made it look like a morningstar instead of a mace.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Chaosvolt <chaosvolt@users.noreply.github.com>