Energy Rework Part 1
Procrastinated long enough. This commit is mainly proof of concept, it makes batteries type: BATTERY instead of type magazine and has them store energy as units::energy instead of charges.

`ammo_consume` is a hack for me to play with, which confirms that power is being drawn correctly when it is called.

The bionic scanner is sort of necessary at the moment for testing purposes.

Next step is to convert energy using functions over from charge usage.