fix: make `scent_modifier` affect your scent equilibrium. (#5323)
Update character_turn.cpp

`scent_modifier` was multiplying the scent value after addition/subtraction. Meaning that if you had a value less than 0 and just created the character/loaded it, the scent map would be 0 and increased to 1, then multiplied by 0.x which is then truncated to 0.

Fixes it so it actually multplies your `norm_scent` value.