Redoes the entire logic for molotov action.
Base molotov has no charges, instead, it transforms into a lit molotov that has charges, that over time depletes it's charges and turns back.

Disallow extinguishing it manually, as this produces molotovs with no max charge but a charge. They don't stack in the UI with other molotovs This issue will still occur if it gets extinguished by water.