add carbondioxide cylinder, soda machine, and soda recipes
ALTERNATIVE:
could have created a magazine like weldgas/oxyacetylene and torches. Would have required an item for each called "gasname_dispenser" which would essentially be a reskinned welding tank. nitrogen was already implemented this way, so we opted to go with that. well, c_hydrogen_gas is implemented like welding tank, so now im split. can we craft with ammo?

we need to work towards an air compressor or atmospheric gas separator from labs. This will allow the player to craft gases using only time? or certain reagents. unfortunately it seems currently a single gas item wont be able to be used in a recipe and as a magazine/ammo type at the same time.