fix: Make the electrosense bionic scanner actually power down (#6162)
* Make the electrosense bionic actually power down

I misunderstood the purpose of the `eff_only` argument being true, which is meant to only trigger the deactivation effect without actually running the check to see if the bionic can actually deactivate or deactivate the bionic.

This is apparently done to simulate involuntary deactivations that have no relation to the other pre-reqs like deactivation cost.

* Update bionics.cpp