feat: compactor machine furniture (#6171)
* Metal compactor furniture

The metal compactor can now be deconstructed into an item that can be placed as furniture wherever and used for the same purpose, making mass item deconstruction a bit easier. Also removed the compactor's sprite config so it doesn't look identical to recycling bins, making it a bit easier to visually identify what it is.

* Apparently-trash-compactors-have-internal-space

* Damn you checks!

* Update tile_config.json

* Double Spaces? What madness