Flail rebalance, new weap reworking
Made the flails smaller and more lightweight, since they're basically intended to be a ball and chain on the end of a spearhead.  Increased their damage too, and gave them the belt_clip flag. Cut the trench club entirely and reflavored the spiked club as a trench club. Renamed spiked club and survivor staff to 'spiked survivor x'.