Energy Rework Part 6.4
Finish all the patching fixes needed to make this damn thing compile.

This is incredibly preliminary, it will fail on loading JSONs because the JSONs largely have not been updated to the new style.

Also, recipe, construction and crafting have not been updated in any way to interface with the new system, so I'll need to get on that for Part 7.

TODO: UPDATE TESTS, update recipe, construction and crafting recipes and how they draw charges/power.