Energy Rework 6.4.2
Tentatively uses assigns in `active_tile_data.cpp` to allow power and the like to use both legacy and "new" methods of defining power.

Migration now checks if contents is not empty before attempting to check if it's battery charges.

Activation requirements are now listed in the game_inventory menu.