fix: Fridgetanks work now (#6360)
Temp considerations to vehicle::process_contents

Hate this so much, but for some reason active item processing queues aren't being added to when it comes to fluids and tanks.

Can't be bothered to fix this at the moment, so here's a patchfix.

Also updates fridgetanks to actually be able to contain things.