more recipe sanity checking, soda syrups
includes cutting as well as boiling, containing, cooking to represent cutting the fruit up for the recipe. (energy drinks have nothing to cut up and therefore dont need it) base time is increased (45min to steep/turn into syrup, additional 30 to cool and bottle) as well as batch time being added to the recipe. the recipes have also been separated into a syrup and an actual soda recipe