feat: (atomic) soda machines, co2 tanks & "suicide" soda drink (#6372)
* add carbondioxide cylinder, soda machine, and soda recipes

ALTERNATIVE:
could have created a magazine like weldgas/oxyacetylene and torches. Would have required an item for each called "gasname_dispenser" which would essentially be a reskinned welding tank. nitrogen was already implemented this way, so we opted to go with that. well, c_hydrogen_gas is implemented like welding tank, so now im split. can we craft with ammo?

we need to work towards an air compressor or atmospheric gas separator from labs. This will allow the player to craft gases using only time? or certain reagents. unfortunately it seems currently a single gas item wont be able to be used in a recipe and as a magazine/ammo type at the same time.

* add soda machine and carbondioxide to spawn groups where food dehydrator and nitrogen are found

* add more soda recipes

* recipe fix

* atomic soda machine & recipes, energy cola rename

* mmm lint

* description rework

* add additional sodas to sizzurp, looks_like tag for liquid

* price sanity check

i think the item descriptions were already fixed? guess we'll find out

* please robot accept the PR

* more recipe sanity checking, soda syrups

includes cutting as well as boiling, containing, cooking to represent cutting the fruit up for the recipe. (energy drinks have nothing to cut up and therefore dont need it) base time is increased (45min to steep/turn into syrup, additional 30 to cool and bottle) as well as batch time being added to the recipe. the recipes have also been separated into a syrup and an actual soda recipe

* style(autofix.ci): automated formatting

* pushed PR too soon. story of my life. add sewer soda & pine soda as well

* style(autofix.ci): automated formatting

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>