feat: scouting electronic multitool visor (#6687)
* What's the Scouter Say?

Added in "scouting visor", an all-in-one electronic multitool that replaces, and is crafted from, a laptop/smartphone/smartwatch, a GPS, a geiger counter, and night vision goggles. Added scouting visor to relevant itemgroups, and adjusted itemgroup numbers for the GPS and 3D printer to make the GPS more common and the 3D printer slightly less.

Also added in recipe for a civilian GPS and military GPS, as well as a recipe to upgrade the scouter to use the military GPS's radius.