feat: Enable a wider variety of Adult Toys (#7170)
* Initial support for sex_toy use action

It doesn't use the charges on activation that it's supposed to, but tbh I kinda think that it's goofy to have that separate anyway

* Bullet Vibe Support

Funnily enough, the (mostly) generic reversion code here could probably also get used for mp3

* Actual bullet vibe item

* Removes charges_per_use from vibrator

It didn't really make sense to be there anyway?

And now it doesn't work with the sex_toy action

* update MN's mana vibrator

* Put the bullet vibe into itemgroups

No crafting recipe or putting it on NPCs yet, though.

* style(autofix.ci): automated formatting

* fix bullet vibes' issues

* Update electronics.json

* Make the bullet vibe trigger every minute

* Update electronics.json

* Update electronics.json

* Add recipe for bullet vibe

Co-Authored-By: RoyalFox <112293514+RoyalFox2140@users.noreply.github.com>

* Make the good feeling last longer

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: RoyalFox <112293514+RoyalFox2140@users.noreply.github.com>