Mapgen now supports batteries
Code now supports the ability for items in item groups to be defined with "energy" array or "energy-min" and "energy-max", and will spawn the item with the default battery assuming that the item group has the "battery" field set to a number above 0.

Overall functions the same way charges used to and magazines currently do. Documentation updated.